# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend service for the Cryptocash Battery on the TON network.",
        "title": "Battery implementation for TON network",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/relayer/dev-send": {
            "post": {
                "description": "TEST ONLY — makes the relayer send TON to an address (deploys the relayer on first send, pays gas). Enabled only when DEV_SIGN=true. Response carries header X-Dev-Route: TO-BE-DELETED.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dev"
                ],
                "summary": "[TO BE DELETED] Dev-only relayer send",
                "parameters": [
                    {
                        "type": "string",
                        "description": "destination address",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "amount in TON (default 0.05)",
                        "name": "amount",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "optional comment",
                        "name": "comment",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.RelayerDevSendResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/ton-proof/check": {
            "post": {
                "description": "Verifies a ton_proof signature for a v5R1 wallet, resolves the user by public key, and returns a session JWT",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Verify ton_proof and issue a JWT",
                "parameters": [
                    {
                        "description": "ton_proof payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/tonproof.Request"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.TonProofCheckResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/ton-proof/payload": {
            "get": {
                "description": "Returns a single-use payload (nonce) that the wallet must sign for ton_proof login",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Generate ton_proof payload",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.PayloadResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/wallet/emulate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Runs static checks, emulation and policy over a message BOC and returns Supported-By-Battery / Allowed-By-Battery / Reject-Reason plus the charge cost",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "battery"
                ],
                "summary": "Emulate a message and return the battery decision",
                "parameters": [
                    {
                        "description": "message boc",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.walletEmulateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DecisionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handler.DecisionResponse": {
            "type": "object",
            "properties": {
                "allowed": {
                    "type": "boolean"
                },
                "charge": {
                    "type": "integer"
                },
                "rejectReason": {
                    "type": "string"
                },
                "supported": {
                    "type": "boolean"
                }
            }
        },
...
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handler.PayloadResponse": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "string"
                }
            }
        },
...
        "handler.RelayerDevSendResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "string"
                },
                "explorer": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "txHash": {
                    "type": "string"
                }
            }
        },
...
        "handler.TonProofCheckResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "expiresAt": {
                    "type": "integer"
                },
                "newUser": {
                    "type": "boolean"
                },
                "publicKey": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "handler.walletEmulateRequest": {
            "type": "object",
            "properties": {
                "boc": {
                    "type": "string"
                },
                "ignoreSignature": {
                    "type": "boolean"
                }
            }
        },
...
        "tonproof.Domain": {
            "type": "object",
            "properties": {
                "length_bytes": {
                    "type": "integer"
                },
                "value": {
                    "type": "string"
                }
            }
        },
        "tonproof.Proof": {
            "type": "object",
            "properties": {
                "domain": {
                    "$ref": "#/definitions/tonproof.Domain"
                },
                "payload": {
                    "type": "string"
                },
                "signature": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                }
            }
        },
        "tonproof.Request": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "network": {
                    "type": "string"
                },
                "proof": {
                    "$ref": "#/definitions/tonproof.Proof"
                },
                "public_key": {
                    "type": "string"
                },
                "wallet_state_init": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "AdminToken": {
            "type": "apiKey",
            "name": "X-Admin-Token",
            "in": "header"
        },
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}